Job api, create job. The output is missing fields and provinces.

+++ src/main/java/vn/unigap/api/repository/JobProvinceRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import vn.unigap.api.entity.jpa.JobProvince;
 
-public interface JobProvinceRepository extends JpaRepository<JobProvince, Long> {
+public interface JobProvinceRepository extends JpaRepository<JobProvince, Integer> {
 }
